Correct type for mallocing in safe_memory file

+++ utility/stack.c
@@ -5,7 +5,7 @@
 
 #define INITIAL_STACK_SIZE 20
 
-stack* create_stack(int element_size){
+stack* create_stack(size_t element_size){
     stack* new_stack = (stack*)safe_malloc(sizeof(stack));
     new_stack->top = -1;
     new_stack->max = INITIAL_STACK_SIZE;
@@ -23,7 +23,7 @@ void* pop(stack* stack){
 void push(stack* stack, void* data){
     if(stack->top == stack->max){
         stack->max *= 2;
-        stack->data = (void**)safe_realloc(stack->data, stack->max);
+        stack->data = (void**)safe_realloc(stack->data, (size_t) stack->max);
     }
     stack->data[++stack->top] = data;
 }
